bug调试

+++ scientificmangement/src/main/java/com/sicnu/pojo/teamExamine/AwardTeamExamine.java
@@ -6,7 +6,7 @@ public class AwardTeamExamine {
     private String user_role;
     private Double contribution;
 
-    public long getAward_id(Integer user_id) {
+    public Integer getAward_id(Integer user_id) {
         return award_id;
     }
 
@@ -14,7 +14,7 @@ public class AwardTeamExamine {
         this.award_id = award_id;
     }
 
-    public long getUser_id() {
+    public Integer getUser_id() {
         return user_id;
     }
