+++ scientificmangement/src/main/java/com/sicnu/pojo/teamExamine/ProjectTeamExamine.java
@@ -3,24 +3,24 @@ package com.sicnu.pojo.teamExamine;
 
 public class ProjectTeamExamine {
 
-  private long pe_id;
-  private long user_id;
+  private Integer pe_id;
+  private Integer user_id;
   private String user_role;
   private String team_role;
 
-  public long getPe_id() {
+  public Integer getPe_id() {
     return pe_id;
   }
 
-  public void setPe_id(long pe_id) {
+  public void setPe_id(Integer pe_id) {
     this.pe_id = pe_id;
   }
 
-  public long getUser_id() {
+  public Integer getUser_id() {
     return user_id;
   }
 
-  public void setUser_id(long user_id) {
+  public void setUser_id(Integer user_id) {
     this.user_id = user_id;
   }
